Move basket status constants onto the Basket class
It was a mistake to have them at module level as it meant it was a pain
to import them when they were needed for filtering.  Having them on
the basket is much nicer as they available whenever the Basket class is.

This commit may cause a few gremlins for people updating to 0.4 but I
think it's a necessary pain at this stage.  It's trivial to fix too.