Re-order URL segments to be more RESTful
Better to use /products/123/delete/ than /products/delete/123/ as the
first is more like a reference to a resource (/products/123/).  Of
course, the truly RESTful way of deleting a product would be to send a
DELETE HTTP request to the /products/123/ but that's not really
practical for a normal web app.

Only a small thing.