Fix for <form> action attribute.
Old behavior is to default to action="." if no action is given. The dot
makes the browser submit to the URL specified, but drops any GET
parameters if the form is POSTed.
A common case of mixing POST and GET parameters is supplying ?_popup=1
to a popup. The view processes the form data in POST as usual, but then
might return a different HTTPResponse if 'popup' was supplied, e.g.
return Javascript to close the popup.

My fix drops specifying 'action' if none is given. This is okay in
HTML5 and HTML3, and defaults to the URL (including GET parameters).
It's required in HTML4, but as far as I know handled gracefully in
all browsers.