Default shipping/billing addresses with one click
Setting a default shipping or billing address used to mean finding the
checkbox in the edit view. It is now possible directly from the
AddressListView via a dropdown.

If an address is a default address for either, it is also marked as such
in the AddressListView.

This commit removes the checkboxes for is_default_for_(billing|shipping)
from the UserAddressForm, hence also hiding them in the address edit views
in the checkout app.