Allow products to be added to basket
This requires changes to `add_product` to allow the chosen stock record
to be passed in.  The various basket properties then use the selected
stockrecord to fetch price information.

This implementation is likely to change to allow tax to be calculated
based on session variables at some point - probably by injecting the
price strategy into the constructor.