Improve reviews display
* The review list and detail page needed some direction, after looking at
amazon and barnes and noble reviews I added a product summary at the top
of the page. This works well as the average rating is also brought
through which adds "write review" button if "you" have now reviewed. The
previous button broke if you clicked on it and you have reviewed.

* Stars have been replaced by font icons - adding more flexibility to
  styling.
* Added a reviews partial