Update migration to include PartnerAddress model.
Without this when running schemamigration --auto on partner a migration to
create PartnerAddress is created.

The reason seems to be that this and the 0009 migration have been created
in parallel on two branches as migration 0009, then this migration has been
renamed to 0010, but not "rebased" to reflect that PartnerAddress now
exists. Therefore, when detecting changes for a schemamigration South
assumes that PartnerAddress doesn't exist.

address.country was added by south when I recreated this migration. I think
this change is not strictly necessary, but I am not 100% sure so I left it
it. Same about the length change and the related_name additions.