Handle Django's new password reset links
Django 1.6 changed the encoding of the user's primary key. See the
comment in oscar/app.py for an overview.

This commit also reverts the URL syntax for all password reset links to
the style in the Django documentation. It is slightly more permissive,
but shouldn't cause any issues.

This commit also changes a bug in the get_password_reset_url function
where it would ignore the passed token generator and always use the
default one.

Fixes #1203