Get weight-based shipping method to return charges again
Setting band.charge was accidentally killed in the refactor to the new
shipping method interface, so a correct charge was never returned. This
wasn't caught because there was no test for testing the actual shipping
charge calculation...

Both issues are remedied now. I'll have a more detailled look at the
shipping method tests when getting rid of upper_charge.