Dashboard search: Fallback to __icontains search for UPC
On Bike Parts Market, UPCs are structured in a way that allows grouping
of products by UPC. In that case, an icontains search on UPC becomes
useful.

This commit introduces a change that if the UPC exacht search doesn't
match, it falls back to an icontains search, potentially showing more
than one match. This mirrors what the title search does already.