Get facets working
This required in passing a variant of our SearchForm. Be careful to not
clash with the 'search_form' context variable that is always passed in.
At least until we've decided what the search field should do; currently
it's entirely separate.

As a nice surprise, sorting results Just Works(tm).

The Haystack search form sways a bit from a regular Django form, and all
we're really interested in is calling .search() on it anyway. So I opted
to not change the view to a FormView, as the form shouldn't really ever
be invalid if not messed with, and there's nothing sensible to do apart
from just showing all results.

I removed load_all() for now and want to do some further investigating
of what a sensible way to proceed is. I suspect load_all would load
the entire (unpaginated!) result set into memory, which might be a bit
much.