Revert "Use string notation for factory boy factories"
This reverts commit f243314e55fefe7211158d97f1ce909398f35f13.

I mistakenly assumed that the string notation would delay the loading of
the model class as it does e.g. when used for ForeignKeys or in the
AUTH_USER_MODEL. But in fact it achieves the opposite, it not just loads
the model, but it requires the entire model registry to be present!

This has been reported in
https://github.com/rbarrois/factory_boy/issues/160

Until then, we need to stick to the old syntax.