Get currency default value to play nice with Django 1.7 migrations
When running oscar on Django 1.7 in an application where you have
modified OSCAR_DEFAULT_CURRENCY the new migration engine detects the
change as something it needs to generate a migration for.

Changed the default value to be a function instead. This prevents the
migration engine from interpreting a settings change as something it
needs to generate a migration for.

Closes #1512.