Fixes for demo site's parent product detail page
The demo site lists child products on a parent product's detail page,
and some of the code needed updating to the new naming scheme. I updated
the child_stock_record.html template to a simpler version of the regular
stock_record.html template.

Also switched from displaying the (generic) child title to displaying
the attributes, and removed the loading of the cache template tags as
they weren't used.

Fixes #1475.