Import models module instead of relying on import_models()
After feedback from @aaugustin, the fix for the Django 1.7 model loading
regression (done in 51a0859) can still interfere with the process. The
clean approach is to emulate importing the models file.

The code was provided from @aaugustin; I merely tested it still fixes
the issue and fiddled with the comments to appease the gods of PEP8.