Switch to using browser-based LESS pre-processor
django-compressor used to trigger processing of the less files during
development. But with the browser-based compiler, an easy replacement is
available. This even lowers the barrier to altering styles, because
getting npm up and running isn't required.

Offline compilation isn't impacted.

Using the most current less version seems to not cause any trouble, so I
removed the version requirement.